Add files via upload
Jarajapupeta is a small village located in the Vizianagaram district of Andhra Pradesh, India. It is known for its rural setting and agriculture-based economy, with farming being the primary livelihood of the local population. The village, like many in the region, has a close-knit community that celebrates traditional festivals and maintains strong cultural practices. Telugu is the primary language spoken here. The village is part of a broader area rich in natural beauty and has a simple, rural lifestyle. Though modest in infrastructure, Jarajapupeta is an example of the vibrant rural communities that contribute to the cultural and agricultural landscape of Vizianagaram.